+++ src/db/db_eathena.c
@@ -244,7 +244,7 @@ int const_ea_load(void * db, int row, int col, char * val) {
         case 0: strnload(record->name, MAX_NAME_SIZE, val);                             break;
         case 1:
             /* constant can be represented as hexadecimal or decimal */
-            record->value = (strlen(val) > 2 && val[0] == '0' && val[1] == 'x') ?
+            record->value = ((int) strlen(val) > 2 && val[0] == '0' && val[1] == 'x') ?
                 convert_integer(val, 16):
                 convert_integer(val, 10);                                               break;
         case 2: record->type = convert_integer(val, 10);                                break;
@@ -296,15 +296,15 @@ int ea_db_init(ea_db_t ** ea, const char * path) {
         goto failed;
     }
 
-    if( SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_ITEM_INSERT, strlen(EA_ITEM_INSERT), &_ea->item_ea_sql_insert, NULL) ||
-        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_MOB_INSERT, strlen(EA_MOB_INSERT), &_ea->mob_ea_sql_insert, NULL) ||
-        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_SKILL_INSERT, strlen(EA_SKILL_INSERT), &_ea->skill_ea_sql_insert, NULL) ||
-        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_PRODUCE_INSERT, strlen(EA_PRODUCE_INSERT), &_ea->produce_ea_sql_insert, NULL) ||
-        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_MERC_INSERT, strlen(EA_MERC_INSERT), &_ea->mercenary_ea_sql_insert, NULL) ||
-        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_PET_INSERT, strlen(EA_PET_INSERT), &_ea->pet_ea_sql_insert, NULL) ||
-        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_ITEM_GROUP_INSERT, strlen(EA_ITEM_GROUP_INSERT), &_ea->item_group_ea_sql_insert, NULL) ||
-        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_ITEM_GROUP_ENTRY_INSERT, strlen(EA_ITEM_GROUP_ENTRY_INSERT), &_ea->item_group_ea_sql_record_insert, NULL) ||
-        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_CONST_INSERT, strlen(EA_CONST_INSERT), &_ea->const_ea_sql_insert, NULL)) {
+    if( SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_ITEM_INSERT, (int) strlen(EA_ITEM_INSERT), &_ea->item_ea_sql_insert, NULL) ||
+        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_MOB_INSERT, (int) strlen(EA_MOB_INSERT), &_ea->mob_ea_sql_insert, NULL) ||
+        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_SKILL_INSERT, (int) strlen(EA_SKILL_INSERT), &_ea->skill_ea_sql_insert, NULL) ||
+        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_PRODUCE_INSERT, (int) strlen(EA_PRODUCE_INSERT), &_ea->produce_ea_sql_insert, NULL) ||
+        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_MERC_INSERT, (int) strlen(EA_MERC_INSERT), &_ea->mercenary_ea_sql_insert, NULL) ||
+        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_PET_INSERT, (int) strlen(EA_PET_INSERT), &_ea->pet_ea_sql_insert, NULL) ||
+        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_ITEM_GROUP_INSERT, (int) strlen(EA_ITEM_GROUP_INSERT), &_ea->item_group_ea_sql_insert, NULL) ||
+        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_ITEM_GROUP_ENTRY_INSERT, (int) strlen(EA_ITEM_GROUP_ENTRY_INSERT), &_ea->item_group_ea_sql_record_insert, NULL) ||
+        SQLITE_OK != sqlite3_prepare_v2(_ea->db, EA_CONST_INSERT, (int) strlen(EA_CONST_INSERT), &_ea->const_ea_sql_insert, NULL)) {
         fprintf(stderr, "[load]: failed prepare sql statments; %s.\n", sqlite3_errmsg(_ea->db));
         goto failed;
     }
@@ -382,8 +382,8 @@ int ea_db_item_load_record(item_ea * items, int size, sqlite3_stmt * sql) {
         item = &items[i];
         if( SQLITE_OK != sqlite3_clear_bindings(sql) ||
             SQLITE_OK != sqlite3_bind_int(sql, 1, item->id) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 2, item->aegis, strlen(item->aegis), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 3, item->eathena, strlen(item->eathena), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 2, item->aegis, (int) strlen(item->aegis), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 3, item->eathena, (int) strlen(item->eathena), SQLITE_STATIC) ||
             SQLITE_OK != sqlite3_bind_int(sql, 4, item->type) ||
             SQLITE_OK != sqlite3_bind_int(sql, 5, item->buy) ||
             SQLITE_OK != sqlite3_bind_int(sql, 6, item->sell) ||
@@ -400,9 +400,9 @@ int ea_db_item_load_record(item_ea * items, int size, sqlite3_stmt * sql) {
             SQLITE_OK != sqlite3_bind_int(sql, 17, item->elv) ||
             SQLITE_OK != sqlite3_bind_int(sql, 18, item->refineable) ||
             SQLITE_OK != sqlite3_bind_int(sql, 19, item->view) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 20, item->script, strlen(item->script), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 21, item->onequip, strlen(item->onequip), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 22, item->onunequip, strlen(item->onunequip), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 20, item->script, (int) strlen(item->script), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 21, item->onequip, (int) strlen(item->onequip), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 22, item->onunequip, (int) strlen(item->onunequip), SQLITE_STATIC) ||
             SQLITE_DONE != sqlite3_step(sql) ||
             SQLITE_OK != sqlite3_reset(sql)) {
             fprintf(stderr, "[load]: failed to add %s to item db.\n", item->aegis);
@@ -443,9 +443,9 @@ int ea_db_mob_load_record(mob_ea * mobs, int size, sqlite3_stmt * sql) {
         mob = &mobs[i];
         if( SQLITE_OK != sqlite3_clear_bindings(sql) ||
             SQLITE_OK != sqlite3_bind_int(sql, 1, mob->id) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 2, mob->sprite, strlen(mob->sprite), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 3, mob->kro, strlen(mob->kro), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 4, mob->iro, strlen(mob->iro), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 2, mob->sprite, (int) strlen(mob->sprite), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 3, mob->kro, (int) strlen(mob->kro), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 4, mob->iro, (int) strlen(mob->iro), SQLITE_STATIC) ||
             SQLITE_OK != sqlite3_bind_int(sql, 5, mob->lv) ||
             SQLITE_OK != sqlite3_bind_int(sql, 6, mob->hp) ||
             SQLITE_OK != sqlite3_bind_int(sql, 7, mob->sp) ||
@@ -541,22 +541,22 @@ int ea_db_skill_load_record(skill_ea * skills, int size, sqlite3_stmt * sql) {
         skill = &skills[i];
         if( SQLITE_OK != sqlite3_clear_bindings(sql) ||
             SQLITE_OK != sqlite3_bind_int(sql, 1, skill->id) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 2, skill->range.str, strlen(skill->range.str), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 2, skill->range.str, (int) strlen(skill->range.str), SQLITE_STATIC) ||
             SQLITE_OK != sqlite3_bind_int(sql, 3, skill->hit) ||
             SQLITE_OK != sqlite3_bind_int(sql, 4, skill->inf) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 5, skill->element.str, strlen(skill->element.str), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 5, skill->element.str, (int) strlen(skill->element.str), SQLITE_STATIC) ||
             SQLITE_OK != sqlite3_bind_int(sql, 6, skill->nk) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 7, skill->splash.str, strlen(skill->splash.str), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 7, skill->splash.str, (int) strlen(skill->splash.str), SQLITE_STATIC) ||
             SQLITE_OK != sqlite3_bind_int(sql, 8, skill->maxlv) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 9, skill->hit_amount.str, strlen(skill->hit_amount.str), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 10, skill->cast_cancel, strlen(skill->cast_cancel), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 9, skill->hit_amount.str, (int) strlen(skill->hit_amount.str), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 10, skill->cast_cancel, (int) strlen(skill->cast_cancel), SQLITE_STATIC) ||
             SQLITE_OK != sqlite3_bind_int(sql, 11, skill->cast_def_reduce_rate) ||
             SQLITE_OK != sqlite3_bind_int(sql, 12, skill->inf2) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 13, skill->maxcount.str, strlen(skill->maxcount.str), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 14, skill->type, strlen(skill->type), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 15, skill->blow_count.str, strlen(skill->blow_count.str), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 16, skill->name, strlen(skill->name), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 17, skill->desc, strlen(skill->desc), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 13, skill->maxcount.str, (int) strlen(skill->maxcount.str), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 14, skill->type, (int) strlen(skill->type), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 15, skill->blow_count.str, (int) strlen(skill->blow_count.str), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 16, skill->name, (int) strlen(skill->name), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 17, skill->desc, (int) strlen(skill->desc), SQLITE_STATIC) ||
             SQLITE_DONE != sqlite3_step(sql) ||
             SQLITE_OK != sqlite3_reset(sql)) {
             fprintf(stderr, "[load]: failed to add %s to skill db.\n", skill->name);
@@ -603,9 +603,9 @@ int ea_db_produce_load_record(produce_ea * produces, int size, sqlite3_stmt * sq
             SQLITE_OK != sqlite3_bind_int(sql, 3, produce->skill_id) ||
             SQLITE_OK != sqlite3_bind_int(sql, 4, produce->skill_lv) ||
             !array_to_string(buf, produce->item_id_req) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 5, buf, strlen(buf), SQLITE_TRANSIENT) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 5, buf, (int) strlen(buf), SQLITE_TRANSIENT) ||
             !array_to_string_cnt(buf, produce->item_amount_req, array_field_cnt(buf) + 1) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 6, buf, strlen(buf), SQLITE_TRANSIENT) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 6, buf, (int) strlen(buf), SQLITE_TRANSIENT) ||
             SQLITE_DONE != sqlite3_step(sql) ||
             SQLITE_OK != sqlite3_reset(sql)) {
             fprintf(stderr, "[load]: failed to add %d to produce db.\n", produce->item_id);
@@ -649,8 +649,8 @@ int ea_db_merc_load_record(mercenary_ea * mercs, int size, sqlite3_stmt * sql) {
         merc = &mercs[i];
         if( SQLITE_OK != sqlite3_clear_bindings(sql) ||
             SQLITE_OK != sqlite3_bind_int(sql, 1, merc->id) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 2, merc->sprite, strlen(merc->sprite), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 3, merc->name, strlen(merc->name), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 2, merc->sprite, (int) strlen(merc->sprite), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 3, merc->name, (int) strlen(merc->name), SQLITE_STATIC) ||
             SQLITE_OK != sqlite3_bind_int(sql, 4, merc->lv) ||
             SQLITE_OK != sqlite3_bind_int(sql, 5, merc->hp) ||
             SQLITE_OK != sqlite3_bind_int(sql, 6, merc->sp) ||
@@ -715,8 +715,8 @@ int ea_db_pet_load_record(pet_ea * pets, int size, sqlite3_stmt * sql) {
         pet = &pets[i];
         if( SQLITE_OK != sqlite3_clear_bindings(sql) ||
             SQLITE_OK != sqlite3_bind_int(sql, 1, pet->mob_id) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 2, pet->name, strlen(pet->name), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 3, pet->jname, strlen(pet->jname), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 2, pet->name, (int) strlen(pet->name), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 3, pet->jname, (int) strlen(pet->jname), SQLITE_STATIC) ||
             SQLITE_OK != sqlite3_bind_int(sql, 4, pet->lure_id) ||
             SQLITE_OK != sqlite3_bind_int(sql, 5, pet->egg_id) ||
             SQLITE_OK != sqlite3_bind_int(sql, 6, pet->equip_id) ||
@@ -734,8 +734,8 @@ int ea_db_pet_load_record(pet_ea * pets, int size, sqlite3_stmt * sql) {
             SQLITE_OK != sqlite3_bind_int(sql, 18, pet->attack_rate) ||
             SQLITE_OK != sqlite3_bind_int(sql, 19, pet->defence_attack_rate) ||
             SQLITE_OK != sqlite3_bind_int(sql, 20, pet->change_target_rate) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 21, pet->pet_script, strlen(pet->pet_script), SQLITE_STATIC) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 22, pet->loyal_script, strlen(pet->loyal_script), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 21, pet->pet_script, (int) strlen(pet->pet_script), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 22, pet->loyal_script, (int) strlen(pet->loyal_script), SQLITE_STATIC) ||
             SQLITE_DONE != sqlite3_step(sql) ||
             SQLITE_OK != sqlite3_reset(sql)) {
             fprintf(stderr, "[load]: failed to add %s to pet db.\n", pet->name);
@@ -875,7 +875,7 @@ int ea_db_const_load_record(const_ea * constants, int size, sqlite3_stmt * sql) 
     for(i = 0; i < size; i++) {
         constant = &constants[i];
         if( SQLITE_OK != sqlite3_clear_bindings(sql) ||
-            SQLITE_OK != sqlite3_bind_text(sql, 1, constant->name, strlen(constant->name), SQLITE_STATIC) ||
+            SQLITE_OK != sqlite3_bind_text(sql, 1, constant->name, (int) strlen(constant->name), SQLITE_STATIC) ||
             SQLITE_OK != sqlite3_bind_int(sql, 2, constant->value) ||
             SQLITE_OK != sqlite3_bind_int(sql, 3, constant->type) ||
             SQLITE_DONE != sqlite3_step(sql) ||
